Comments and fixes due to changed working environment
The compilers used functions found on my previous dev machine which are
lacking in my current one. These are makearbitrarymap and util.map.
The former was a hack around Matlab's atrocious associative array
handling and the latter is a more friendly user interface to arrayfun
and cellfun. makearbitrarymap was replaced by a slightly less hacky
version: previously it defined an arbitrary function calling eval while
now a seperate function is used to get values. util.map was found in a
user made package and abused Matlab's exception handling mechanism, for
portability with Octave, it has been removed and replaced with standard
cellfun and arrayfun